+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/computercraft/peripherals/InformativeRegistryPeripheral.kt
@@ -4,9 +4,14 @@ import dan200.computercraft.api.lua.LuaException
 import dan200.computercraft.api.lua.LuaFunction
 import dan200.computercraft.api.lua.MethodResult
 import net.minecraft.resources.ResourceLocation
+import net.minecraft.world.entity.LivingEntity
+import net.minecraft.world.level.block.Blocks
+import net.minecraft.world.level.material.Fluids
 import site.siredvin.broccolium.modules.platform.PlatformRegistries
 import site.siredvin.peripheralworks.common.blockentity.InformativeRegistryBlockEntity
 import site.siredvin.peripheralworks.common.configuration.PeripheralWorksConfig
+import site.siredvin.peripheralworks.common.setup.Items
+import site.siredvin.peripheralworks.xplat.ModPlatform
 import site.siredvin.tweakium.modules.peripheral.OwnedPeripheral
 import site.siredvin.tweakium.modules.peripheral.owner.BlockEntityPeripheralOwner
 import site.siredvin.tweakium.modules.peripheral.representation.LuaRepresentation
@@ -31,6 +36,40 @@ class InformativeRegistryPeripheral(
 
         init {
             addList(
+                "mods",
+                "Minecraft mods",
+                {
+                    MethodResult.of(ModPlatform.modList.filter { !PeripheralWorksConfig.informativeRegistryModBlocklist.contains(it) })
+                },
+                {
+                    if (PeripheralWorksConfig.informativeRegistryModBlocklist.contains(it)) {
+                        return@addList MethodResult.of(null)
+                    }
+                    MethodResult.of(ModPlatform.getModInformation(it))
+                },
+            )
+            addList(
+                "entity",
+                "Minecraft living entities",
+                {
+                    MethodResult.of(
+                        PlatformRegistries.ENTITY_TYPES.keySet().filter {
+                            PlatformRegistries.ENTITY_TYPES.get(it)?.baseClass?.isAssignableFrom(LivingEntity::class.java) ?: false
+                        }.map(ResourceLocation::toString),
+                    )
+                },
+                {
+                    @Suppress("DEPRECATION", "KotlinRedundantDiagnosticSuppress")
+                    val entityType = PlatformRegistries.ENTITY_TYPES.get(ResourceLocation(it))
+                    val data: MutableMap<String, Any> = mutableMapOf()
+                    data["category"] = entityType.category.name
+                    data["type"] = entityType.description.string
+                    data["defaultLootTable"] = entityType.defaultLootTable.toString()
+                    data["registryID"] = PlatformRegistries.ENTITY_TYPES.getId(entityType)
+                    return@addList MethodResult.of(data)
+                },
+            )
+            addList(
                 "item",
                 "Minecraft items",
                 {
@@ -39,6 +78,9 @@ class InformativeRegistryPeripheral(
                 {
                     @Suppress("DEPRECATION", "KotlinRedundantDiagnosticSuppress")
                     val item = PlatformRegistries.ITEMS.get(ResourceLocation(it))
+                    if (item == net.minecraft.world.item.Items.AIR) {
+                        return@addList MethodResult.of(null)
+                    }
                     val base = LuaRepresentation.forItem(item)
                     base["registryID"] = PlatformRegistries.ITEMS.getId(item)
                     return@addList MethodResult.of(base)
@@ -54,6 +96,9 @@ class InformativeRegistryPeripheral(
                 {
                     @Suppress("DEPRECATION", "KotlinRedundantDiagnosticSuppress")
                     val blockState = PlatformRegistries.BLOCKS.get(ResourceLocation(it)).defaultBlockState()
+                    if (blockState.`is`(Blocks.AIR)) {
+                        return@addList MethodResult.of(null)
+                    }
                     val base = LuaRepresentation.forBlockState(blockState)
                     base["registryID"] = PlatformRegistries.BLOCKS.getId(blockState.block)
                     return@addList MethodResult.of(base)
@@ -69,6 +114,9 @@ class InformativeRegistryPeripheral(
                 {
                     @Suppress("DEPRECATION", "KotlinRedundantDiagnosticSuppress")
                     val fluid = PlatformRegistries.FLUIDS.get(ResourceLocation(it))
+                    if (fluid == Fluids.EMPTY) {
+                        return@addList MethodResult.of(null)
+                    }
                     val base = LuaRepresentation.forFluid(fluid)
                     base["registryID"] = PlatformRegistries.FLUIDS.getId(fluid)
                     return@addList MethodResult.of(base)