+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/xplat/ModInnerPlatform.kt
@@ -5,4 +5,6 @@ import site.siredvin.tweakium.modules.platform.api.InnerComputerBasePlatform
 
 interface ModInnerPlatform : InnerComputerBasePlatform {
     val commonEnergy: EnergyUnit
+    val modList: List<String>
+    fun getModInformation(mod: String): Map<String, Any>?
 }