+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/xplat/ModPlatform.kt
@@ -25,4 +25,8 @@ object ModPlatform : ComputerBasePlatform() {
 
     val commonEnergy: EnergyUnit
         get() = baseInnerPlatform.commonEnergy
+
+    val modList: List<String>
+        get() = baseInnerPlatform.modList
+    fun getModInformation(mod: String): Map<String, Any>? = baseInnerPlatform.getModInformation(mod)
 }